Paste raw link when cursor is inside a markdown link (#2)
* Paste raw link when cursor is inside a markdown link

i.e. when pasting inside [Anchor text](<pasting here>)

* Fix inability to paste hashtags on Chrome

Standalone hashtags are valid URL objects, but not a valid URL
pattern for querying tabs. This adds a check and doesn't consider
standalone hashtags as valid URL in the content script.

* Fix inability to paste URLs with leading or trailing whitespace

Leading/trailing whitespace is ignored by the URL constructor but
is invalid when querying tabs